refactor: rename and reorganize onboarding stage definitions
