+++ src/lib/client/cutscene/definitions/stages/getting-started/help.ts
