+++ src/lib/client/cutscene/definitions/stages/getting-started/personalize.ts
